o Change various attach arg and softc datatypes to ints. o Use explicit typecasts when interfacing with MRG data
Should fix port-mac68k/6839.  Patch supplied by
Frederick Bruckman <fb@enteract.com>

+++ sys/arch/mac68k/dev/kbdvar.h
@@ -1,4 +1,4 @@
-/*	$Id: kbdvar.h,v 1.1 1998/10/23 01:16:24 ender Exp $	*/
+/*	$Id: kbdvar.h,v 1.2 1999/02/11 06:41:09 ender Exp $	*/
 
 /*
  * Copyright (C) 1998	Colin Wood
@@ -42,9 +42,9 @@ struct kbd_softc {
 	struct  device  sc_dev;
 
 	/* ADB info */
-	u_int8_t        origaddr;       /* ADB device type (ADBADDR_KBD) */
-	u_int8_t        adbaddr;        /* current ADB address */
-	u_int8_t        handler_id;     /* type of keyboard */
+	int	        origaddr;       /* ADB device type (ADBADDR_KBD) */
+	int	        adbaddr;        /* current ADB address */
+	int	        handler_id;     /* type of keyboard */
 
 	u_int8_t	sc_leds;	/* current LED state */
 };
